requirements.txt
Pillow~=9.5.0
pyinstaller~=5.10.1